refactor(chat): extract chat_sessions.py (session CRUD, history, feedback, delete)
Verbatim move of create_session, get_session, get_session_for_user,
list_sessions, get_history, get_history_paginated, _resolve_session,
update_message_feedback, delete_session. Re-imported into
app.services.chat; api-layer imports and patch targets unchanged.
819 tests pass.
